// lib/api/generated/models/meetingReviewResponse.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Heymoa API
 * Heymoa 서버 REST API
 * OpenAPI spec version: 1.0.0
 */
import type { MeetingReviewResponseData } from "./meetingReviewResponseData";
import type { MeetingReviewResponseError } from "./meetingReviewResponseError";

export interface MeetingReviewResponse {
  /** 성공 응답 데이터 */
  data: MeetingReviewResponseData;
  /** 요청 성공 여부 */
  success: boolean;
  /** 성공 시 null인 오류 정보 */
  error: MeetingReviewResponseError;
}